Reorganize stuff.
